Remove useless border style check from wxVListBox::Create()
The comparison could never be true as wxDEFAULT is not one of
wxBORDER_XXX values and so is not included in wxBORDER_MASK and so was
useless -- and just results in -Wtautological-compare from gcc 8.

Simply remove this check, 750777261cd4890a5bad8d7c6b121cbe18ade6d1 added
it apparently to use themed border by default for wxVListBox, but this
was done in a better way in 4abbd614fa2dc5ae85d53f26af9f7079ecf02427.