Fix search button height in generic wxSearchCtrl
Make this button of exactly the bitmap size instead of using the full
control height for it, even if/when the bitmap is smaller.

This fixes some random junk being shown in the bottom two pixels of the
search button sometimes: this was due to only redrawing the part covered
by the bitmap in wxSearchButton::OnPaint(), leaving the rest of the
button unpainted (because wxSearchButton uses wxBG_STYLE_PAINT).