Fix drawing of static box with empty label under MSW
The top part of the box border was not drawn after the changes of
695f40a78414b26f652b6be596dfd0341b79bb7b which decreased the top border
size too much in the case of empty label.

Fix this and also try to improve the code a little bit by introducing
another symbolic constant in MSW code and not using the constant from
the base class code, as like this we can at least keep all the related
constants together -- even if they're still as magic as ever.

Closes #18191.